pyproject.toml
[project]
name = "unify-2-1-dm-dqa"
version = "0.1.0"
description = "Unify 2.1 Data Quality Assessment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cli-core>=2.74.0",
    "azure-core>=1.34.0",
    "azure-devops>=7.1.0b4",
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.10.0",
    "azure-mgmt-devtestlabs>=9.0.0",
    "azure-mgmt-resource>=24.0.0",
    "azure-mgmt-storage>=23.0.1",
    "azure-storage-blob>=12.25.1",
    "azure-synapse-artifacts>=0.20.0",
    "chispa>=0.11.1",
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "icecream>=2.1.4",
    "loguru>=0.7.3",
    "msal>=1.32.3",
    "msal-extensions>=1.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "psutil>=7.0.0",
    "pydantic==2.6.3",
    "pyspark==3.4.0",
    "pytest>=8.4.1",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=6.2.1",
    "pytest-fixture-classes>=1.0.3",
    "pytest-html>=4.1.1",
    "pytest-icdiff>=0.9",
    "pytest-json>=0.4.0",
    "pytest-loguru>=0.4.0",
    "pytest-pretty>=1.3.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "snoop>=0.6.0",
]

[tool.ruff]
target-version = "py312"
respect-gitignore = true
line-length = 240
indent-width = 4
output-format = "full"
show-fixes = true
exclude = [ ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d", ".pyenv", ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv", ".vscode", "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules", "site-packages", "venv",]

[tool.ruff.lint]
select = [ "E4", "E7", "E9", "F",]
ignore = [ "E712", "F811", "F403", "F405",]
fixable = [ "ALL",]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-line-length = "dynamic"
